requirements.txt
piexif==1.0.13
Pillow==5.0.0
